MNIST/requirements.txt
gradio>=4.0,<5.0
numpy>=1.24
pillow>=9.0
pandas>=2.0
